+++ src/Conflux.js
@@ -22,7 +22,7 @@ class Conflux {
    * @param [options.logger] {Object} - Logger object with 'info' and 'error' method.
    * @example
    * > const { Conflux } = require('js-conflux-sdk');
-   * > const conflux = new Conflux({url:'http://test.confluxrpc.org'});
+   * > const conflux = new Conflux({url:'http://test.confluxrpc.org', chainId: 1});
    *
    * @example
    * > const conflux = new Conflux({
@@ -35,6 +35,8 @@ class Conflux {
     defaultGasPrice,
     defaultGasRatio = 1.1,
     defaultStorageRatio = 1.1,
+    chainId,
+    useHexAddressInParameter = false,
     ...rest
   } = {}) {
     /**
@@ -90,6 +92,13 @@ class Conflux {
 
     this.sendRawTransaction = this._decoratePendingTransaction(this.sendRawTransaction);
     this.sendTransaction = this._decoratePendingTransaction(this.sendTransaction);
+
+    if (chainId) {
+      this.chainId = chainId;
+      this.wallet.setNetId(chainId);
+    }
+
+    this.useHexAddressInParameter = useHexAddressInParameter;
   }
 
   _decoratePendingTransaction(func) {
@@ -97,6 +106,21 @@ class Conflux {
     return function (...args) {
       return new PendingTransaction(conflux, func.bind(this), args);
     };
+  }
+
+  _formatAddress(address) {
+    if (!this.chainId) {
+      console.warn('Conflux address: chainId is not set properly, please set it');
+    }
+    return this.useHexAddressInParameter ? format.hexAddress(address) : format.address(address, this.chainId);
+  }
+
+  _formatCallTx(options) {
+    return format.callTxAdvance(this.chainId, this.useHexAddressInParameter)(options);
+  }
+
+  _formatGetLogs(options) {
+    return format.getLogsAdvance(this.chainId, this.useHexAddressInParameter)(options);
   }
 
   /**
@@ -153,6 +177,15 @@ class Conflux {
   }
 
   // --------------------------------------------------------------------------
+  /**
+   * Update conflux chainId from RPC
+   */
+  async updateChainId() {
+    const { chainId } = await this.getStatus();
+    this.chainId = chainId;
+    this.wallet.setNetId(chainId);
+  }
+
   /**
    * Get node client version
    *
@@ -275,20 +308,20 @@ class Conflux {
    * - admin `string`: admin of the account.
    *
    * @example
-   > await conflux.getAccount('0x1c1e72f0c37968557b3d85a3f32747792798bbde');
+   > await conflux.getAccount('cfxtest:00e1wwrgsdwpgnbv7p2u7wt78xwjf65vvsdz67gx0u');
    {
       accumulatedInterestReturn: 0n,
       balance: 824812401057514588670n,
       collateralForStorage: 174187500000000000000n,
       nonce: 1449n,
       stakingBalance: 0n,
-      admin: '0x0000000000000000000000000000000000000000',
+      admin: 'CFXTEST:TYPE.NULL:0000000000000000000000000000000000W5PHD2EJ',
       codeHash: '0xc5d2460186f7233c927e7db2dcc703c0e500b653ca82273b7bfad8045d85a470'
    }
    */
   async getAccount(address, epochNumber) {
     const result = await this.provider.call('cfx_getAccount',
-      format.address(address),
+      this._formatAddress(address),
       format.epochNumber.$or(undefined)(epochNumber),
     );
     return format.account(result);
@@ -302,12 +335,12 @@ class Conflux {
    * @return {Promise<BigInt>} The balance in Drip.
    *
    * @example
-   * > await conflux.getBalance("0x1c1e72f0c37968557b3d85a3f32747792798bbde");
+   * > await conflux.getBalance("cfxtest:00e1wwrgsdwpgnbv7p2u7wt78xwjf65vvsdz67gx0u");
    824812401057514588670n
    */
   async getBalance(address, epochNumber) {
     const result = await this.provider.call('cfx_getBalance',
-      format.address(address),
+      this._formatAddress(address),
       format.epochNumber.$or(undefined)(epochNumber),
     );
     return format.bigUInt(result);
@@ -321,12 +354,12 @@ class Conflux {
    * @return {Promise<BigInt>} The staking balance in Drip.
    *
    * @example
-   * > await conflux.getStakingBalance('0x1c1e72f0c37968557b3d85a3f32747792798bbde', 'latest_state');
+   * > await conflux.getStakingBalance('cfxtest:00e1wwrgsdwpgnbv7p2u7wt78xwjf65vvsdz67gx0u', 'latest_state');
    0n
    */
   async getStakingBalance(address, epochNumber) {
     const result = await this.provider.call('cfx_getStakingBalance',
-      format.address(address),
+      this._formatAddress(address),
       format.epochNumber.$or(undefined)(epochNumber),
     );
     return format.bigUInt(result);
@@ -340,12 +373,12 @@ class Conflux {
    * @return {Promise<BigInt>} The next nonce should be used by given address.
    *
    * @example
-   * > await conflux.getNextNonce("0x1c1e72f0c37968557b3d85a3f32747792798bbde");
+   * > await conflux.getNextNonce("cfxtest:00e1wwrgsdwpgnbv7p2u7wt78xwjf65vvsdz67gx0u");
    1449n
    */
   async getNextNonce(address, epochNumber) {
     const result = await this.provider.call('cfx_getNextNonce',
-      format.address(address),
+      this._formatAddress(address),
       format.epochNumber.$or(undefined)(epochNumber),
     );
     return format.bigUInt(result);
@@ -359,12 +392,12 @@ class Conflux {
    * @return {Promise<string>} Address to admin, or `null` if the contract does not exist.
    *
    * @example
-   * > conflux.getAdmin('0x8e2f2e68eb75bb8b18caafe9607242d4748f8d98')
-   "0x1c1e72f0c37968557b3d85a3f32747792798bbde"
+   * > conflux.getAdmin('cfxtest:0272ybk8xduvr2sstaryjs3j8ba793wdk0s62tr4gj')
+   "CFXTEST:TYPE.USER:00E1WWRGSDWPGNBV7P2U7WT78XWJF65VVSDZ67GX0U"
    */
   async getAdmin(address, epochNumber) {
     return this.provider.call('cfx_getAdmin',
-      format.address(address),
+      this._formatAddress(address),
       format.epochNumber.$or(undefined)(epochNumber),
     );
   }
@@ -381,7 +414,7 @@ class Conflux {
    */
   async getVoteList(address, epochNumber) {
     const result = await this.provider.call('cfx_getVoteList',
-      format.address(address),
+      this._formatAddress(address),
       format.epochNumber.$or(undefined)(epochNumber),
     );
     return format.voteList(result);
@@ -399,7 +432,7 @@ class Conflux {
    */
   async getDepositList(address, epochNumber) {
     const result = await this.provider.call('cfx_getDepositList',
-      format.address(address),
+      this._formatAddress(address),
       format.epochNumber.$or(undefined)(epochNumber),
     );
     return format.depositList(result);
@@ -477,14 +510,14 @@ class Conflux {
       baseReward: 6993700000000000000n,
       totalReward: 6993700031741486703n,
       txFee: 0n,
-      author: '0x1f323dccb24606b061db9e3a1277b8db99f1c1b2',
+      author: 'CFXTEST:TYPE.USER:00FK4FECP930DC31VEF3M4KRR3DTKWE1P8SKN4FE6W',
       blockHash: '0x73cd891aea310e2c0b8644de91746c7353cebfffb780126bc06101b20689c893'
     },
    {
       baseReward: 6997200000000000000n,
       totalReward: 6997200031760371742n,
       txFee: 3000000n,
-      author: '0x1f323dccb24606b061db9e3a1277b8db99f1c1b2',
+      author: 'CFXTEST:TYPE.USER:00FK4FECP930DC31VEF3M4KRR3DTKWE1P8SKN4FE6W',
       blockHash: '0xaf4136d04e9e2cc470703251ec46f5913ab7955d526feed43771705e89c77390'
     }
    ]
@@ -556,7 +589,7 @@ class Conflux {
       deferredReceiptsRoot: '0x09f8709ea9f344a810811a373b30861568f5686e649d6177fd92ea2db7477508',
       deferredStateRoot: '0x50c0fcbc5bafa7d1dba7b19c87629830106a6be8d0adf505cdc656bb43535d69',
       hash: '0xaf4136d04e9e2cc470703251ec46f5913ab7955d526feed43771705e89c77390',
-      miner: '0x1f323dccb24606b061db9e3a1277b8db99f1c1b2',
+      miner: 'CFXTEST:TYPE.USER:00FK4FECP930DC31VEF3M4KRR3DTKWE1P8SKN4FE6W',
       nonce: '0x17d86f2f6',
       parentHash: '0xc8a412b4b77b48d61f694975f032d109f26bb0f9fc02e4b221d67a382fab386b',
       powQuality: '0x5a0f86a6f4',
@@ -651,11 +684,11 @@ class Conflux {
       blockHash: '0xaf4136d04e9e2cc470703251ec46f5913ab7955d526feed43771705e89c77390',
       contractCreated: null,
       data: '0xfebe49090000000000000000000000000000000000000000000000000000000000000000000000000000000000000000162788589c8e386863f217faef78840919fb2854',
-      from: '0x108b8b1333523a79ac363d8f41805e81b085d55d',
+      from: 'CFXTEST:TYPE.USER:00FK4FECP930DC31VEF3M4KRR3DTKWE1P8SKN4FE6W',
       hash: '0xbf7110474779ba2404433ef39a24cb5b277186ef1e6cb199b0b60907b029a1ce',
       r: '0x495da01ae9f445847022a8bc7df0198577ba75f88b26699f61afb435bb9c50bc',
       s: '0x2291051b1c53db1d6bfe2fb29be1bf512d063e726dc6b98aaf0f2259b7456be0',
-      to: '0x83bf953c8b687f0d1b8d2243a3e0654ec1f70d1b'
+      to: 'CFXTEST:TYPE.USER:00FK4FECP930DC31VEF3M4KRR3DTKWE1P8SKN4FE6W'
     }
    */
   async getTransactionByHash(transactionHash) {
@@ -699,11 +732,11 @@ class Conflux {
       gasFee: 1500000n,
       blockHash: '0xaf4136d04e9e2cc470703251ec46f5913ab7955d526feed43771705e89c77390',
       contractCreated: null,
-      from: '0x108b8b1333523a79ac363d8f41805e81b085d55d',
+      from: 'CFXTEST:TYPE.USER:0088R2SK6D93MYDC6SYSYGC0BU0V11ENBM1SY9PC7E',
       logs: [],
       logsBloom: '0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000',
       stateRoot: '0xd6a7c2c14cb0d1233010acca98e114db5a10e0b94803d23b01a6777b7fd3b2fd',
-      to: '0x83bf953c8b687f0d1b8d2243a3e0654ec1f70d1b',
+      to: 'CFXTEST:TYPE.CONTRACT:021VZ59WHDM7Y38VHMH478Z0CN7C3XSD3CS5HCXXDZ',
       transactionHash: '0xbf7110474779ba2404433ef39a24cb5b277186ef1e6cb199b0b60907b029a1ce',
       txExecErrorMsg: null,
       gasCoveredBySponsor: false,
@@ -830,11 +863,11 @@ class Conflux {
     "blockHash": null,
     "contractCreated": null,
     "data": "0x",
-    "from": "0x1bd9e9be525ab967e633bcdaeac8bd5723ed4d6b",
+    "from": "CFXTEST:TYPE.USER:00DXKUDYA9DBJTZ66EYDNUP8RNBJ7VADDCG5CDSGXW",
     "hash": "0xb2ba6cca35f0af99a9601d09ee19c1949d8130312550e3f5413c520c6d828f88",
     "r": "0x245a1a86ae405eb72c1eaf98f5e22baa326fcf8262abad2c4a3e5bdcf2e912b5",
     "s": "0x4df8058887a4dd8aaf60208accb3e57292a50ff06a117df6e54f7f56176248c0",
-    "to": "0x1bd9e9be525ab967e633bcdaeac8bd5723ed4d6b"
+    "to": "CFXTEST:TYPE.USER:00DXKUDYA9DBJTZ66EYDNUP8RNBJ7VADDCG5CDSGXW"
    }
 
    * @example
@@ -853,11 +886,11 @@ class Conflux {
     "blockHash": "0xdb2d2d438dcdee8d61c6f495bd363b1afb68cb0fdff16582c08450a9ca487852",
     "contractCreated": null,
     "data": "0x",
-    "from": "0x1bd9e9be525ab967e633bcdaeac8bd5723ed4d6b",
+    "from": "CFXTEST:TYPE.USER:00DXKUDYA9DBJTZ66EYDNUP8RNBJ7VADDCG5CDSGXW",
     "hash": "0xb2ba6cca35f0af99a9601d09ee19c1949d8130312550e3f5413c520c6d828f88",
     "r": "0x245a1a86ae405eb72c1eaf98f5e22baa326fcf8262abad2c4a3e5bdcf2e912b5",
     "s": "0x4df8058887a4dd8aaf60208accb3e57292a50ff06a117df6e54f7f56176248c0",
-    "to": "0x1bd9e9be525ab967e633bcdaeac8bd5723ed4d6b"
+    "to": "CFXTEST:TYPE.USER:00DXKUDYA9DBJTZ66EYDNUP8RNBJ7VADDCG5CDSGXW"
    }
 
    * @example
@@ -870,11 +903,11 @@ class Conflux {
     "gasFee": 21000000000000n,
     "blockHash": "0xdb2d2d438dcdee8d61c6f495bd363b1afb68cb0fdff16582c08450a9ca487852",
     "contractCreated": null,
-    "from": "0x1bd9e9be525ab967e633bcdaeac8bd5723ed4d6b",
+    "from": "CFXTEST:TYPE.USER:00DXKUDYA9DBJTZ66EYDNUP8RNBJ7VADDCG5CDSGXW",
     "logs": [],
     "logsBloom": "0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
     "stateRoot": "0x510d680cdbf60d34bcd987b3bf9925449c0839a7381dc8fd8222d2c7ee96122d",
-    "to": "0x1bd9e9be525ab967e633bcdaeac8bd5723ed4d6b",
+    "to": "CFXTEST:TYPE.USER:00DXKUDYA9DBJTZ66EYDNUP8RNBJ7VADDCG5CDSGXW",
     "transactionHash": "0xb2ba6cca35f0af99a9601d09ee19c1949d8130312550e3f5413c520c6d828f88"
    }
 
@@ -888,11 +921,11 @@ class Conflux {
     "gasFee": 21000000000000n,
     "blockHash": "0xdb2d2d438dcdee8d61c6f495bd363b1afb68cb0fdff16582c08450a9ca487852",
     "contractCreated": null,
-    "from": "0x1bd9e9be525ab967e633bcdaeac8bd5723ed4d6b",
+    "from": "CFXTEST:TYPE.USER:00DXKUDYA9DBJTZ66EYDNUP8RNBJ7VADDCG5CDSGXW",
     "logs": [],
     "logsBloom": "0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
     "stateRoot": "0x510d680cdbf60d34bcd987b3bf9925449c0839a7381dc8fd8222d2c7ee96122d",
-    "to": "0x1bd9e9be525ab967e633bcdaeac8bd5723ed4d6b",
+    "to": "CFXTEST:TYPE.USER:00DXKUDYA9DBJTZ66EYDNUP8RNBJ7VADDCG5CDSGXW",
     "transactionHash": "0xb2ba6cca35f0af99a9601d09ee19c1949d8130312550e3f5413c520c6d828f88"
    }
    */
@@ -903,7 +936,7 @@ class Conflux {
     }
 
     return this.provider.call('cfx_sendTransaction',
-      format.callTx(options),
+      this._formatCallTx(options),
       password,
     );
   }
@@ -917,12 +950,12 @@ class Conflux {
    * @return {Promise<string>} Byte code of contract, or 0x if the contract does not exist.
    *
    * @example
-   * > await conflux.getCode('0xb385b84f08161f92a195953b980c8939679e906a');
+   * > await conflux.getCode('cfxtest:021sbe2f10b1z4n1jpakr60ch4wpf7mgd8ak9064p5');
    "0x6080604052348015600f57600080fd5b506004361060325760003560e01c806306661abd1460375780638..."
    */
   async getCode(address, epochNumber) {
     return this.provider.call('cfx_getCode',
-      format.address(address),
+      this._formatAddress(address),
       format.epochNumber.$or(undefined)(epochNumber),
     );
   }
@@ -936,12 +969,12 @@ class Conflux {
    * @return {Promise<string|null>} Storage entry of given query, or null if the it does not exist.
    *
    * @example
-   * > await conflux.getStorageAt('0x866aca87ff33a0ae05d2164b3d999a804f583222', '0x6661e9d6d8b923d5bbaab1b96e1dd51ff6ea2a93520fdc9eb75d059238b8c5e9')
+   * > await conflux.getStorageAt('cfxtest:0236njm7zwtu1bg5u8b4pfctka04yp1j48rju7dww2', '0x6661e9d6d8b923d5bbaab1b96e1dd51ff6ea2a93520fdc9eb75d059238b8c5e9')
    "0x000000000000000000000000000000000000000000000000000000000000162e"
    */
   async getStorageAt(address, position, epochNumber) {
     return this.provider.call('cfx_getStorageAt',
-      format.address(address),
+      this._formatAddress(address),
       format.hex64(position),
       format.epochNumber.$or(undefined)(epochNumber),
     );
@@ -958,7 +991,7 @@ class Conflux {
    * - snapshot `string`: storage root in the snapshot.
    *
    * @example
-   * > await conflux.getStorageRoot('0x866aca87ff33a0ae05d2164b3d999a804f583222')
+   * > await conflux.getStorageRoot('cfxtest:0236njm7zwtu1bg5u8b4pfctka04yp1j48rju7dww2')
    {
       "delta": "0xc5d2460186f7233c927e7db2dcc703c0e500b653ca82273b7bfad8045d85a470",
       "intermediate": "0xc5d2460186f7233c927e7db2dcc703c0e500b653ca82273b7bfad8045d85a470",
@@ -967,7 +1000,7 @@ class Conflux {
    */
   async getStorageRoot(address, epochNumber) {
     return this.provider.call('cfx_getStorageRoot',
-      format.address(address),
+      this._formatAddress(address),
       format.epochNumber.$or(undefined)(epochNumber),
     );
   }
@@ -985,18 +1018,18 @@ class Conflux {
    * - sponsorForGas `string`: the address of the gas sponsor.
    *
    * @example
-   * > await conflux.getSponsorInfo('0x8e2f2e68eb75bb8b18caafe9607242d4748f8d98')
+   * > await conflux.getSponsorInfo('cfxtest:0272ybk8xduvr2sstaryjs3j8ba793wdk0s62tr4gj')
    {
       sponsorBalanceForCollateral: 410625000000000000000n,
       sponsorBalanceForGas: 9999999993626232440n,
       sponsorGasBound: 10000000000n,
-      sponsorForCollateral: '0x8d5adbcaf5714924830591586f05302bf87f74bd',
-      sponsorForGas: '0x8d5adbcaf5714924830591586f05302bf87f74bd'
+      sponsorForCollateral: 'CFXTEST:TYPE.CONTRACT:026NNPYAYNSMJ9430P8NGVS560NZGZVMRMHXYVM0BK',
+      sponsorForGas: 'CFXTEST:TYPE.CONTRACT:026NNPYAYNSMJ9430P8NGVS560NZGZVMRMHXYVM0BK'
    }
    */
   async getSponsorInfo(address, epochNumber) {
     const result = await this.provider.call('cfx_getSponsorInfo',
-      format.address(address),
+      this._formatAddress(address),
       format.epochNumber.$or(undefined)(epochNumber),
     );
     return format.sponsorInfo(result);
@@ -1010,12 +1043,12 @@ class Conflux {
    * @return {Promise<BigInt>} - The collateral storage in Byte.
    *
    * @example
-   * > await conflux.getCollateralForStorage('0x8e2f2e68eb75bb8b18caafe9607242d4748f8d98')
+   * > await conflux.getCollateralForStorage('cfxtest:0272ybk8xduvr2sstaryjs3j8ba793wdk0s62tr4gj')
    89375000000000000000n
    */
   async getCollateralForStorage(address, epochNumber) {
     const result = await this.provider.call('cfx_getCollateralForStorage',
-      format.address(address),
+      this._formatAddress(address),
       format.epochNumber.$or(undefined)(epochNumber),
     );
     return format.bigUInt(result);
@@ -1031,7 +1064,7 @@ class Conflux {
   async call(options, epochNumber) {
     try {
       return await this.provider.call('cfx_call',
-        format.callTx(options),
+        this._formatCallTx(options),
         format.epochNumber.$or(undefined)(epochNumber),
       );
     } catch (e) {
@@ -1052,7 +1085,7 @@ class Conflux {
   async estimateGasAndCollateral(options, epochNumber) {
     try {
       const result = await this.provider.call('cfx_estimateGasAndCollateral',
-        format.callTx(options),
+        this._formatCallTx(options),
         format.epochNumber.$or(undefined)(epochNumber),
       );
       return format.estimate(result);
@@ -1084,7 +1117,7 @@ class Conflux {
    *
    * @example
    * > await conflux.getLogs({
-      address: '0x8e2f2e68eb75bb8b18caafe9607242d4748f8d98',
+      address: 'cfxtest:0272ybk8xduvr2sstaryjs3j8ba793wdk0s62tr4gj',
       fromEpoch: 39802,
       toEpoch: 39802,
       limit: 1,
@@ -1096,7 +1129,7 @@ class Conflux {
       logIndex: 2,
       transactionIndex: 0,
       transactionLogIndex: 2,
-      address: '0x8e2f2e68eb75bb8b18caafe9607242d4748f8d98',
+      address: 'CFXTEST:TYPE.CONTRACT:0272YBK8XDUVR2SSTARYJS3J8BA793WDK0S62TR4GJ',
       blockHash: '0xca00158a2a508170278d5bdc5ca258b6698306dd8c30fdba32266222c79e57e6',
       data: '0x',
       topics: [
@@ -1114,7 +1147,7 @@ class Conflux {
       throw new Error('OverrideError, do not use `blockHashes` with `fromEpoch` or `toEpoch`, cause only `blockHashes` will take effect');
     }
 
-    const result = await this.provider.call('cfx_getLogs', format.getLogs(options));
+    const result = await this.provider.call('cfx_getLogs', this._formatGetLogs(options));
 
     return format.logs(result);
   }
@@ -1136,10 +1169,10 @@ class Conflux {
                     {
                         "action": {
                             "callType": "call",
-                            "from": "0x19c742cec42b9e4eff3b84cdedcde2f58a36f44f",
+                            "from": "CFXTEST:TYPE.USER:00CWEGPESGNTWKRZ7E2CVVEDWBUSMDRM9W7KY3YE3R",
                             "gas": "311592",
                             "input": "0x",
-                            "to": "0x84980a94d94f54ac335109393c08c866a21b1b0e",
+                            "to": "CFXTEST:TYPE.CONTRACT:0229G2MMV57N9B1KA44KJF08T1KA46SV1SBG5W9ASV",
                             "value": "0"
                         },
                         "type": "call"
@@ -1150,7 +1183,7 @@ class Conflux {
                 "traces": [
                     {
                         "action": {
-                            "from": "0x1bdd8e198e78a36a1819c06d683ab1f89d2b006d",
+                            "from": "CFXTEST:TYPE.USER:00DXV3GTHTWA6UGS3706UU1UP7W9UAS0DMUW6P42A1",
                             "gas": "83962",
                             "init": "0x",
                             "value": "0"
@@ -1257,7 +1290,7 @@ class Conflux {
       deferredReceiptsRoot: '0x7ae0d5716513206755b6f7c95272b79dbc225759b6e17727e19c2f15c3166bda',
       deferredStateRoot: '0x3cf5deba77c8aa9072f1e972d6a97db487a0ce88455f371eb8ac8fa77321cb9d',
       hash: '0x194675173abbc5aab50326136008774eea1a289e6722c973dfed12b703ee5f2a',
-      miner: '0x189121b8f0cdfef0b56eb22d9cb76c97b9c7cfbc',
+      miner: 'CFXTEST:TYPE.USER:00C928DSY36ZXW5NDUT2V75RDJBVKHYFRGYHY8JY0U',
       nonce: '0x799d35f695950fd6',
       parentHash: '0x4af3cf8cb358e75acad282ffa4b578b6211ea9eeb7cf87c282f120d8a1c809df',
       powQuality: '0xe7ac17feab',
@@ -1297,7 +1330,7 @@ class Conflux {
      logIndex: 0,
      transactionIndex: 0,
      transactionLogIndex: 0,
-     address: '0x84ed30d7ddc5ff82ac271ae4e7add5a8b22a8d71',
+     address: 'CFXTEST:TYPE.CONTRACT:022EUC6RVR2ZZ0NC4WDE9TXDUPMB4AMDE4ZCUXV4YS',
      blockHash: '0xc02689eea6a507250838463c13e6b633479e2757dfb7e9b2593d5c31b54adb63',
      data: '0x0000000000000000000000000000000000000000000000000000000000000001',
      topics: [
@@ -1313,7 +1346,7 @@ class Conflux {
    { revertTo: 568231 }
    */
   async subscribeLogs({ address, topics } = {}) {
-    const id = await this.subscribe('logs', format.getLogs({ address, topics }));
+    const id = await this.subscribe('logs', this._formatGetLogs({ address, topics }));
 
     const subscription = new Subscription(id);
     this.provider.on(id, data => {