+++ src/wallet/PrivateKeyAccount.js
@@ -2,12 +2,14 @@ const { assert } = require('../util');
 const format = require('../util/format');
 const sign = require('../util/sign');
 const Account = require('./Account');
+const CONST = require('../CONST');
 
 class PrivateKeyAccount extends Account {
   /**
    * Create a new PrivateKeyAccount with random privateKey.
    *
    * @param [entropy] {string|Buffer} - Entropy of random account
+   * @param [netId=1029] {Integer} - network id of account
    * @return {PrivateKeyAccount}
    *
    * @example
@@ -15,31 +17,31 @@ class PrivateKeyAccount extends Account {
    PrivateKeyAccount {
       privateKey: '0xd28edbdb7bbe75787b84c5f525f47666a3274bb06561581f00839645f3c26f66',
       publicKey: '0xc42b53ae2ef95fee489948d33df391c4a9da31b7a3e29cf772c24eb42f74e94ab3bfe00bf29a239c17786a5b921853b7c5344d36694db43aa849e401f91566a5',
-      address: '0x1cecb4a2922b7007e236daf0c797de6e55496e84'
+      address: 'cfxtest:00eetd52j8nr01z26vdf1hwrvtr5ajbeggmy2sw5sp'
     }
    * > PrivateKeyAccount.random() // gen a different account from above
    PrivateKeyAccount {
       privateKey: '0x1b67150f56f49556ef7e3899024d83c125d84990d311ec08fa98aa1433bc0f53',
       publicKey: '0xd442207828ffd4dad918fea0d75d42dbea1fe5e3789c00a82e18ce8229714eae3f70b12f2f1abd795ad3e5c52a5a597289eb5096548438c233431f498b47b9a6',
-      address: '0x16c25691aadc3363f5862d264072584f3ebf4613'
+      address: 'cfxtest:00bc4nmhnbe36szngspjcg3jb17kxfu62cn2b9n83s'
     }
    * > PrivateKeyAccount.random('0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff');
    PrivateKeyAccount {
       privateKey: '0x1d41e006afd28ea339922d8ab4be93154a14d4f1b6d0ad4e7aabf807e7536a5f',
       publicKey: '0x4c07c75d3fdc5b1d6afef6ec374b0eaac86bcaa771a1d536bc4ce6f111b1c60e414b370e4cf31bf7770ae6818a3518c485398a43857d9053153f6eb4f5644a90',
-      address: '0x113d49784c80d6f8fdbc0bef5a5ab0d9c9fee520'
+      address: 'cfxtest:008kujbs9j0ddy7xrg5yypjup3cwkzr5402rbwgdte'
     }
    * > PrivateKeyAccount.random('0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff');
    * // gen a different account from above, even use same entropy
    PrivateKeyAccount {
       privateKey: '0x5a34ff3318674c33209ce856218890e9a6ee3811e8a51e3094ed1e6a94bf58ef',
       publicKey: '0xe530d77c3ed6115cb46ba79821085bf67d2a7a8c808c1d52dec03fd7a82e569c2136dba84b21d40f46d90484722b21a9d5a8038495adf93f2eed564ababa2422',
-      address: '0x1f63fcef4aaa88c03cbb5c9fb34be69dee65d0a8'
+      address: 'cfxtest:00fp7z7f9an8hg1wrde9zcubwueywtegn0wngx3sdd'
     }
    */
-  static random(entropy) {
+  static random(entropy, netId = CONST.MAINNET_ID) {
     const privateKeyBuffer = sign.randomPrivateKey(entropy === undefined ? undefined : format.hexBuffer(entropy));
-    return new this(privateKeyBuffer);
+    return new this(privateKeyBuffer, netId);
   }
 
   /**
@@ -47,6 +49,7 @@ class PrivateKeyAccount extends Account {
    *
    * @param keystore {object} - Keystore version 3 object.
    * @param password {string|Buffer} - Password for keystore to decrypt with.
+   * @param [netId=1029] {Integer} - Network id of account
    * @return {PrivateKeyAccount}
    *
    * @example
@@ -75,33 +78,35 @@ class PrivateKeyAccount extends Account {
     privateKey: '0x0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef'
   }
    */
-  static decrypt(keystore, password) {
+  static decrypt(keystore, password, netId = CONST.MAINNET_ID) {
     const privateKeyBuffer = sign.decrypt(keystore, password);
-    return new this(privateKeyBuffer);
+    return new this(privateKeyBuffer, netId);
   }
 
   /**
    * Create a account by privateKey.
    *
    * @param privateKey {string|Buffer} - Private key of account
+   * @param [netId] {Integer} - Network id of account
    * @return {PrivateKeyAccount}
    *
    * @example
    * > new PrivateKeyAccount('0x0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef')
    PrivateKeyAccount {
-    address: '0x1cad0b19bb29d4674531d6f115237e16afce377c',
+    address: 'cfxtest:00eau2strcmx8tu567bf2593fsbazkhrfgw83tdrex',
     publicKey: '0x4646ae5047316b4230d0086c8acec687f00b1cd9d1dc634f6cb358ac0a9a8ffffe77b4dd0a4bfb95851f3b7355c781dd60f8418fc8a65d14907aff47c903a559',
     privateKey: '0x0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef'
   }
    */
-  constructor(privateKey) {
+  constructor(privateKey, netId = CONST.MAINNET_ID) {
     const privateKeyBuffer = format.hexBuffer(privateKey);
     const publicKeyBuffer = sign.privateKeyToPublicKey(privateKeyBuffer);
     const addressBuffer = sign.publicKeyToAddress(publicKeyBuffer);
 
-    super(format.address(addressBuffer));
+    super(format.address(addressBuffer, netId));
     this.publicKey = format.publicKey(publicKeyBuffer);
     this.privateKey = format.privateKey(privateKeyBuffer);
+    this.netId = netId;
   }
 
   /**
@@ -137,7 +142,7 @@ class PrivateKeyAccount extends Account {
     })
 
    Transaction {
-      from: '0x1cad0b19bb29d4674531d6f115237e16afce377c',
+      from: 'cfxtest:00eau2strcmx8tu567bf2593fsbazkhrfgw83tdrex',
       nonce: 0,
       gasPrice: 100,
       gas: 10000,
@@ -181,7 +186,7 @@ class PrivateKeyAccount extends Account {
    */
   async signMessage(options) {
     const message = await super.signMessage(options);
-    message.sign(this.privateKey); // sign will cover r,s,v fields
+    message.sign(this.privateKey, this.netId); // sign will cover r,s,v fields
 
     assert(message.from === this.address, {
       message: 'Invalid sign message.from',