+++ test/util/format.test.js
@@ -1,5 +1,5 @@
 const Big = require('big.js');
-const { format } = require('../../src');
+const { format, CONST } = require('../../src');
 const JSBI = require('../../src/util/jsbi');
 
 const HEX_64 = '0x0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef';
@@ -156,14 +156,14 @@ test('hex40', () => {
   expect(format.checksumAddress('0x1B716c51381e76900EBAA7999A488511A4E1fD0A'))
     .toEqual('0x1B716c51381e76900EBAA7999A488511A4E1fD0a');
 
-  expect(format.address('0x1b716c51381e76900ebaa7999a488511a4e1fd0a'))
-    .toEqual('0x1b716c51381e76900ebaa7999a488511a4e1fd0a');
-  expect(format.address('0X1B716C51381E76900EBAA7999A488511A4E1FD0A'))
-    .toEqual('0x1b716c51381e76900ebaa7999a488511a4e1fd0a');
-  expect(format.address('0x1B716c51381e76900EBAA7999A488511A4E1fD0a'))
-    .toEqual('0x1b716c51381e76900ebaa7999a488511a4e1fd0a');
-  expect(() => format.address('0x1B716c51381e76900EBAA7999A488511A4E1fD0A'))
-    .toThrow('checksum error');
+  expect(format.address('0x1b716c51381e76900ebaa7999a488511a4e1fd0a', CONST.TESTNET_ID))
+    .toEqual('cfxtest:00dr2v2h70f7d40eraktk6j8gm8u9sfx185y4xy1w1');
+  expect(format.address('0X1B716C51381E76900EBAA7999A488511A4E1FD0A', CONST.TESTNET_ID))
+    .toEqual('cfxtest:00dr2v2h70f7d40eraktk6j8gm8u9sfx185y4xy1w1');
+  expect(format.address('0x1B716c51381e76900EBAA7999A488511A4E1fD0a', CONST.TESTNET_ID))
+    .toEqual('cfxtest:00dr2v2h70f7d40eraktk6j8gm8u9sfx185y4xy1w1');
+//   expect(() => format.address('0x1B716c51381e76900EBAA7999A488511A4E1fD0A'))
+//     .toThrow('checksum error');
 });
 
 test('hex64', () => {
@@ -228,4 +228,17 @@ test('keccak256', () => {
   expect(format.keccak256([0x42])).toEqual('0x1f675bff07515f5df96737194ea945c36c41e7b4fcef307b7cd4d0e602a69111');
   expect(format.keccak256([0x42])).not.toEqual(format.keccak256('0x42'));
   expect(format.keccak256([0x42])).toEqual(format.keccak256(format.hexBuffer('0x42')));
+});
+
+test('address', () => {
+  expect(() => format.address(undefined)).toThrow('not match "hex"');
+  expect(() => format.address()).toThrow('not match "hex"');
+  expect(() => format.address(null)).toThrow('not match "hex40"');
+  expect(format.address('0x0123456789012345678901234567890123456789', 1)).toEqual('cfxtest:000j6hb7h40j6hb7h40j6hb7h40j6hb7h4mrga5njc');
+  expect(format.address('cfxtest:000j6hb7h40j6hb7h40j6hb7h40j6hb7h4mrga5njc')).toEqual('cfxtest:000j6hb7h40j6hb7h40j6hb7h40j6hb7h4mrga5njc');
+  expect(format.address(Buffer.from('0123456789012345678901234567890123456789', 'hex'), 1)).toEqual('cfxtest:000j6hb7h40j6hb7h40j6hb7h40j6hb7h4mrga5njc');
+
+  expect(() => format.address('0x0123456789012345678')).toThrow('not match "hex40"');
+  expect(() => format.address('cfx:123')).toThrow('not match "hex"');
+  expect(() => format.address('cfx123')).toThrow('not match "hex"');
 });