+++ test/util/sign.test.js
@@ -1,5 +1,5 @@
 const lodash = require('lodash');
-const { format, sign } = require('../../src');
+const { format, sign, CONST } = require('../../src');
 
 const {
   checksumAddress,
@@ -18,7 +18,7 @@ const {
 
 const KEY = '0x0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef';
 const PUBLIC = '0x4646ae5047316b4230d0086c8acec687f00b1cd9d1dc634f6cb358ac0a9a8ffffe77b4dd0a4bfb95851f3b7355c781dd60f8418fc8a65d14907aff47c903a559';
-const ADDRESS = '0x1cad0b19bb29d4674531d6f115237e16afce377c';
+const ADDRESS = 'cfxtest:00eau2strcmx8tu567bf2593fsbazkhrfgw83tdrex';
 const PASSWORD = 'password';
 
 test('checksumAddress', () => {
@@ -58,12 +58,12 @@ test('privateKeyToPublicKey', () => {
 });
 
 test('publicKeyToAddress', () => {
-  const address = format.address(publicKeyToAddress(format.hexBuffer(PUBLIC)));
+  const address = format.address(publicKeyToAddress(format.hexBuffer(PUBLIC)), CONST.TESTNET_ID);
   expect(address).toEqual(ADDRESS);
 });
 
 test('privateKeyToAddress', () => {
-  const address = format.address(privateKeyToAddress(format.hexBuffer(KEY)));
+  const address = format.address(privateKeyToAddress(format.hexBuffer(KEY)), CONST.TESTNET_ID);
   expect(address).toEqual(ADDRESS);
 });
 
@@ -103,6 +103,6 @@ test('ecdsaSign and ecdsaRecover', () => {
   expect(Number.isInteger(v)).toEqual(true);
 
   const publicKey = ecdsaRecover(hash, { r, s, v });
-  const address = format.hex(publicKeyToAddress(publicKey));
+  const address = format.address(publicKeyToAddress(publicKey), CONST.TESTNET_ID);
   expect(address).toEqual(ADDRESS);
 });